Update from Head

+++ Sources/Squall.swift
@@ -261,6 +261,6 @@ fileprivate func Gaussian(mu: Double, sigma: Double, A: Double, B: Double) -> Do
 }
 
 fileprivate func Gaussian(mu: Float, sigma: Float, A: Float, B: Float) -> Float {
-    let X = sqrt(-2*log(A))*cos(2*Float(Double.pi)*B)
+    let X = sqrt(-2*log(A))*cos(2*Float.pi*B)
     return X*sigma + mu
 }
